api/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
google-genai==1.21.1
Pillow==11.0.0
Werkzeug==3.1.3
requests==2.32.3
python-dotenv==1.0.1